third_party/WebKit/LayoutTests/vibration/vibration-iframe-expected.txt

CONSOLE ERROR: line 301: Uncaught TypeError: Cannot read property 'then' of undefined
CONSOLE ERROR: line 301: Uncaught TypeError: Cannot read property 'then' of undefined
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught TypeError: Cannot read property 'then' of undefined
PASS Iframe reload cancels vibration started by it before. 
PASS Iframe destroy cancels vibration started by it before. 
Harness: the test ran to completion.
